Shopping Cart Added

+++ src/main/java/com/books/dto/book/GetBookByCategory.java
@@ -41,7 +41,7 @@ public class GetBookByCategory {
     private int quantity;
 
     @NotNull(message = "Price must be provided")
-    private float price;
+    private Double price;
 
     @NotBlank(message = "Category name must be blank")
     private String categoryName;
